A basic smoke test for cabal-to-dhall.
This entails splitting cabal-to-dhall/Main.hs into a thin driver, and
moving the actual meat of the converter to lib/CabalToDhall.hs.
