+++ dhall-to-cabal.cabal
@@ -81,11 +81,13 @@ source-repository head
 library
     exposed-modules:
         DhallToCabal
+        CabalToDhall
     build-depends:
         Cabal ^>=2.0,
         base ^>=4.10,
         bytestring ^>=0.10,
         containers ^>=0.5,
+        contravariant ^>=1.4,
         dhall ^>=1.12.0,
         formatting ^>=6.3.1,
         hashable ^>=1.2.6.1,
@@ -123,15 +125,9 @@ executable  cabal-to-dhall
     main-is: Main.hs
     scope: public
     build-depends:
-        Cabal ^>=2.0,
         base ^>=4.10,
-        contravariant ^>=1.4,
-        dhall ^>=1.12.0,
-        hashable ^>=1.2.6.1,
         dhall-to-cabal -any,
-        insert-ordered-containers ^>=0.2.1.0,
         optparse-applicative ^>=0.13.2 || ^>=0.14,
-        prettyprinter ^>=1.2.0.1,
         text ^>=1.2
     default-language: Haskell2010
     other-extensions: NamedFieldPuns
