+++ dhall-to-cabal.dhall
@@ -146,6 +146,7 @@ in    prelude.utils.GitHub-project
                   , deps.base
                   , deps.bytestring
                   , deps.containers
+                  , deps.contravariant
                   , deps.dhall
                   , deps.formatting
                   , deps.hashable
@@ -159,7 +160,9 @@ in    prelude.utils.GitHub-project
                     prelude.defaults.CompilerOptions
                   ⫽ { GHC = [ "-Wall", "-fno-warn-name-shadowing" ] }
               , exposed-modules =
-                  [ "DhallToCabal" ]
+                  [ "DhallToCabal"
+                  , "CabalToDhall"
+                  ]
               , hs-source-dirs =
                   [ "lib" ]
               , other-extensions =
@@ -209,15 +212,9 @@ in    prelude.utils.GitHub-project
             "cabal-to-dhall"
             (   prelude.defaults.Executable
               ⫽ { build-depends =
-                    [ deps.Cabal
-                    , deps.base
-                    , deps.contravariant
-                    , deps.dhall
-                    , deps.hashable
+                    [ deps.base
                     , deps.dhall-to-cabal
-                    , deps.insert-ordered-containers
                     , deps.optparse-applicative
-                    , deps.prettyprinter
                     , deps.text
                     ]
                 , hs-source-dirs =
