+++ golden-tests/cabal-to-dhall/simple.cabal
@@ -1,0 +1,3 @@
+cabal-version: 2.0
+name: test
+version: 1.0
