+++ golden-tests/dhall-to-cabal/conditions.cabal
