+++ golden-tests/dhall-to-cabal/dhall-to-cabal.cabal
