+++ golden-tests/dhall-to-cabal/empty-package.cabal
