+++ golden-tests/dhall-to-cabal/lens.cabal
