+++ golden-tests/dhall-to-cabal/nested-conditions.cabal
