+++ golden-tests/dhall-to-cabal/nested-conditions.dhall
