+++ golden-tests/dhall-to-cabal/str-sig.cabal
